requirements.txt
openai>=1.50.0
streamlit>=1.39.0
requests>=2.32.0
python-dotenv>=1.0.0
